Separated visualization from measurments

+++ lib/objects.h
@@ -13,8 +13,8 @@ namespace Objects
 	{
 	public:
 		Cell() {}
-		Cell(size_t n_x, size_t n_y) : x(n_x), y(n_y) {}
-		size_t x,y;
+		Cell(int n_x, int n_y) : x(n_x), y(n_y) {}
+		int x,y;
 	};
 
 	class Rect
